+++ R/audio_filesize.R
@@ -11,6 +11,8 @@
 #' @param duration Duration of recording
 #' @param duration.unit One of seconds, minutes, hours, days
 #' @param output.unit "human", "bits" or  "bytes"
+#' @param units Units used when output.unit is "human", either "decimal" or
+#'   "binary". See humanBytes().
 #' @export
 #' @return The size of the audio file in the specified unit
 #' @examples
@@ -20,7 +22,7 @@
 #' # One year of stereo 24-bit audio sampled at 96kHz
 #' audio_filesize(samp.rate=96000, bit.depth=24, channels=2, duration=1, duration.unit="years")
 #'
-audio_filesize <- function(samp.rate=44100, bit.depth=16, channels=1, duration=1, duration.unit="seconds", output.unit="bits") {
+audio_filesize <- function(samp.rate=44100, bit.depth=16, channels=1, duration=1, duration.unit="seconds", output.unit="bits", units="decimal") {
   duration <- convert2seconds(duration, duration.unit)
   bits <- samp.rate * bit.depth * duration * channels
   if (output.unit == "bits") {
@@ -28,7 +30,7 @@ audio_filesize <- function(samp.rate=44100, bit.depth=16, channels=1, duration=1
   } else if (output.unit == "bytes") {
     return (convert2bytes(bits, "bits"))
   } else if (output.unit == "human") {
-    return (humanBytes(convert2bytes(bits, "bits")))
+    return (humanBytes(convert2bytes(bits, "bits"), units=units))
   } else {
     stop(paste("Unknown output unit:", output.unit))
   }